hello-world-agent/pyproject.toml
[project]
name = "hello-world-agent"
version = "0.1.0"
description = "Hello-world agent with MCP tools and Lakebase memory"
requires-python = ">=3.11"
dependencies = [
    "mlflow>=3.0.0",
    "langchain-core>=0.3.0",
    "databricks-langchain[memory]>=0.16.0",
    "langgraph>=0.3.0",
    "databricks-sdk>=0.40.0",
    "databricks-mcp>=0.3.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.38",
]

[project.scripts]
start-app = "start_server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
